[BAD-1241]-Create and Dev Validate EMR 5.9 Shim against Pentaho 8.1 (Master)
