--- shims/emr56/assemblies/emr56-shim/src/main/resources/config.properties
@@ -1,18 +1,0 @@
-# Friendly name for this configuration
-name=Amazon EMR 5.6.x
-
-# Comma-separated list of directories and files to make available to this
-# configuration. Any resources found here will overwrite ones in lib/.
-classpath=
-
-# Comma-separated list of paths that contain native libraries to load. These
-# could be added to LD_LIBRARY_PATH or set with -Djava.library.path instead.
-library.path=
-
-# Comma-separated list of classes or package names to explicitly ignore when
-# loading classes from the resources within this Hadoop configuration directory
-# or the classpath property
-# e.g.: org.apache.commons.log,org.apache.log4j
-# Note, the two packages above are automatically included for all configurations
-ignore.classes=com.ctc.wstx.stax
-
