--- shims/emr56/assemblies/pom.xml
@@ -1,21 +1,0 @@
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.pentaho</groupId>
-    <artifactId>pentaho-hadoop-shims-emr56-reactor</artifactId>
-    <version>8.1.0.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>pentaho-hadoop-shims-emr56-assemblies-reactor</artifactId>
-  <packaging>pom</packaging>
-
-  <modules>
-    <module>client</module>
-    <module>pmr</module>
-    <module>default</module>
-    <module>emr56-shim</module>
-  </modules>
-</project>
