--- shims/emr56/impl/src/main/resources/META-INF/services/org.pentaho.hadoop.shim.common.authorization.NoOpHadoopAuthorizationService
@@ -1,1 +1,0 @@
-org.pentaho.hadoop.shim.emr56.authorization.ShimNoOpHadoopAuthorizationService
