+++ shims/emr59/assemblies/client/src/assembly/assembly.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<assembly>
+  <id>plugin</id>
+  <baseDirectory>emr59</baseDirectory>
+  <formats>
+    <format>zip</format>
+  </formats>
+  <dependencySets>
+    <dependencySet>
+      <outputDirectory>lib/client</outputDirectory>
+      <useTransitiveDependencies>true</useTransitiveDependencies>
+      <useTransitiveFiltering>false</useTransitiveFiltering>
+      <includes>
+	    <include>org.xerial.snappy:snappy-java</include>
+        <include>dk.brics.automaton:automaton</include>
+        <include>com.google.code.gson:gson</include>
+        <include>joda-time:joda-time</include>
+        <include>commons-configuration:commons-configuration</include>
+      </includes>
+      <excludes>
+        <exclude>xml-apis:*</exclude>
+        <exclude>commons-collections:*</exclude>
+        <exclude>javax.xml.stream:*</exclude>
+        <exclude>org.apache.calcite:*</exclude>
+        <exclude>*:tests:*</exclude>
+      </excludes>
+    </dependencySet>
+    <dependencySet>
+      <outputDirectory>lib/client</outputDirectory>
+      <useTransitiveDependencies>true</useTransitiveDependencies>
+      <useTransitiveFiltering>true</useTransitiveFiltering>
+      <includes>
+        <include>org.apache.hadoop:hadoop-common</include>
+        <include>org.apache.hadoop:hadoop-auth</include>
+        <include>org.apache.zookeeper:zookeeper</include>
+        <include>org.apache.hadoop:hadoop-aws</include>
+        <include>org.apache.hadoop:hadoop-mapreduce-client-app</include>
+        <include>org.apache.hadoop:hadoop-mapreduce-client-core</include>
+        <include>org.apache.hadoop:hadoop-mapreduce-client-hs</include>
+        <include>org.apache.hadoop:hadoop-mapreduce-client-common</include>
+        <include>org.apache.hadoop:hadoop-mapreduce-client-jobclient</include>
+        <include>org.apache.hadoop:hadoop-mapreduce-client-shuffle</include>
+      </includes>
+      <excludes>
+        <exclude>xml-apis:*</exclude>
+        <exclude>*:stax-api:*</exclude>
+        <exclude>commons-collections:*</exclude>
+        <exclude>javax.xml.stream:*</exclude>
+        <exclude>org.apache.calcite:*</exclude>
+        <exclude>*:tests:*</exclude>
+      </excludes>
+    </dependencySet>
+  </dependencySets>
+</assembly>
