+++ shims/emr59/assemblies/default/src/assembly/assembly.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<assembly>
+  <id>plugin</id>
+  <baseDirectory>emr59</baseDirectory>
+  <formats>
+    <format>zip</format>
+  </formats>
+  <dependencySets>
+    <dependencySet>
+      <outputDirectory>lib</outputDirectory>
+      <useTransitiveDependencies>true</useTransitiveDependencies>
+      <useTransitiveFiltering>false</useTransitiveFiltering>
+      <includes>
+		<include>org.xerial.snappy:snappy-java</include>
+        <include>org.apache.hive:hive-jdbc</include>
+        <include>org.apache.hive:hive-exec</include>
+        <include>org.apache.hive:hive-service</include>
+        <include>org.apache.hive:hive-metastore</include>
+        <include>org.apache.thrift:libfb303</include>
+        <include>org.apache.thrift:libthrift</include>
+        <include>org.apache.oozie:oozie-core</include>
+        <include>org.apache.oozie:oozie-client</include>
+        <include>org.apache.sqoop:sqoop</include>
+        <include>org.apache.pig:pig</include>
+        <include>org.python:jython</include>
+        <include>pentaho:pentaho-hdfs-vfs</include>
+        <include>org.apache.parquet:parquet-hadoop-bundle</include>
+        <include>org.apache.avro:avro</include>
+      </includes>
+      <excludes>
+        <exclude>javax.xml.stream:*</exclude>
+        <exclude>org.apache.calcite:*</exclude>
+        <exclude>*:tests:*</exclude>
+      </excludes>
+    </dependencySet>
+  </dependencySets>
+</assembly>
