+++ shims/emr59/assemblies/emr59-shim/src/main/resources/core-site.xml
