+++ shims/emr59/impl/src/main/java/org/apache/hadoop/hive/jdbc/HiveDriver.java
