+++ shims/emr59/impl/src/main/java/org/pentaho/hadoop/shim/emr59/HadoopShim.java
@@ -1,0 +1,60 @@
+/*******************************************************************************
+ *
+ * Pentaho Big Data
+ *
+ * Copyright (C) 2002-2017 by Hitachi Vantara : http://www.pentaho.com
+ *
+ *******************************************************************************
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ ******************************************************************************/
+
+package org.pentaho.hadoop.shim.emr59;
+
+import org.pentaho.hadoop.shim.HadoopConfiguration;
+import org.pentaho.hadoop.shim.HadoopConfigurationFileSystemManager;
+import org.pentaho.hadoop.shim.common.FileSystemProxyV2;
+import org.pentaho.hadoop.shim.common.HadoopShimImpl;
+import org.pentaho.hadoop.shim.common.ShimUtils;
+import org.pentaho.hdfs.vfs.HDFSFileProvider;
+
+import java.io.IOException;
+
+public class HadoopShim extends HadoopShimImpl {
+
+  static {
+    JDBC_DRIVER_MAP.put( "hive2", org.apache.hive.jdbc.HiveDriver.class );
+  }
+
+  @Override
+  public void onLoad( HadoopConfiguration config, HadoopConfigurationFileSystemManager fsm ) throws Exception {
+    super.onLoad( config, fsm );
+    if ( !fsm.hasProvider( "s3n" ) ) {
+      fsm.addProvider( config, "s3n", config.getIdentifier(), new HDFSFileProvider() );
+    }
+  }
+
+  @Override public org.pentaho.hadoop.shim.api.fs.FileSystem getFileSystem(
+    org.pentaho.hadoop.shim.api.Configuration conf ) throws IOException {
+    // Set the context class loader when instantiating the configuration
+    // since org.apache.hadoop.conf.Configuration uses it to load resources
+    ClassLoader cl = Thread.currentThread().getContextClassLoader();
+    Thread.currentThread().setContextClassLoader( getClass().getClassLoader() );
+    try {
+      return new FileSystemProxyV2( ShimUtils.asConfiguration( conf ) );
+    } finally {
+      Thread.currentThread().setContextClassLoader( cl );
+    }
+  }
+}
