+++ shims/emr59/impl/src/main/resources/META-INF/services/org.pentaho.hadoop.shim.common.authorization.NoOpHadoopAuthorizationService
@@ -1,0 +1,1 @@
+org.pentaho.hadoop.shim.emr59.authorization.ShimNoOpHadoopAuthorizationService
